+++ cloud/module/consul/configer.go
@@ -1,0 +1,5 @@
+package consul
+
+type ConfigerClient struct {
+	client *Client
+}
