+++ cloud/registry/registry.go
@@ -10,7 +10,7 @@ import (
 // from https://github.com/go-kratos/kratos/blob/main/contrib/registry/consul/registry.go
 type Registrar interface {
 	// Register the registration.
-	Register(ctx context.Context, service *cloud.RemoteService) error
+	Register(ctx context.Context, service *cloud.RemoteService, opts ...registryOption) error
 	// Deregister the registration.
 	Deregister(ctx context.Context, service *cloud.RemoteService) error
 }
